% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\docType{methods}
\name{acca-method}
\alias{acca-method}
\alias{print.acca}
\alias{summary.acca}
\title{Methods for acca objects}
\usage{
\method{print}{acca}(x, ...)

\method{summary}{acca}(object, ...)
}
\arguments{
\item{x}{an \code{acca} object}

\item{...}{additional arguments to be passed}

\item{object}{an \code{acca}}
}
\description{
Methods for \code{acca} objects. 
The following are methods to extract and manipulate relevant information from
a \code{acca} object.
}
