% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acca_package.R
\docType{package}
\name{acca-package}
\alias{acca-package}
\title{a canonical correlation analysis}
\description{
a canonical correlation analysis
}
\examples{
set.seed(1)
X=matrix(rnorm(500),100,5)
Y=matrix(rnorm(700),100,7)
Z=matrix(rnorm(200),100,2)
mod=cc(X,Y,Z)
mod$cor
mod$scores$corr.X.xscores
mod$prop_expl_var
ggbiplot2(mod)

}
\author{
Livio Finos
}
