% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cc_inference.R
\name{cc_inference}
\alias{cc_inference}
\title{cc_inference}
\usage{
cc_inference(mod, nperms = 100, alpha = 0.5, numb_cc = NULL)
}
\arguments{
\item{mod}{an \code{acca} object.}

\item{nperms}{(\code{100} by default)}

\item{alpha}{(\code{0.5} by default).}

\item{numb_cc}{(\code{=NULL} by default).}
}
\description{
compute p-values
}
